#160 - add tests for TemplateVariableSemanticTokensProvider
